Switched to millisecond timestamps for the items, and added documentation to indicate that this may all be out-of-order anyway, based on the characteristics of the internal storage, clock drift between processors etc.
For comparison, Comsos DB does not provide any guarantees about the in-order nature of its second-resolution timestamps.

+++ Solutions/Marain.Workflow.Abstractions/Marain/Workflows/IWorkflowInstanceChangeLogReader.cs
@@ -15,10 +15,13 @@ namespace Marain.Workflows
         /// Gets a page of log entries for a particular workflow instance ID.
         /// </summary>
         /// <param name="workflowInstanceId">The ID of the workflow instance for which to get log entries.</param>
-        /// <param name="startingTimestamp">The initial sequence number with which to start enumerating entries.</param>
+        /// <param name="startingTimestamp">The unix timestamp with millisecond resolution, with which to start enumerating entries.</param>
         /// <param name="maxItems">The maximum number of items per page. Note that fewer than this number may be returned in any page.</param>
         /// <param name="continuationToken">The continuation token supplied from a previous page of results.</param>
         /// <returns>A page of log entires for the given workflow instance.</returns>
-        Task<WorkflowInstanceLogPage> GetLogEntriesAsync(string workflowInstanceId, int? startingTimestamp = null, int maxItems = 25, string continuationToken = null);
+        /// <remarks>
+        /// <para>Note that there is no absolute guarantee of ordering of the timestamps in the log entries; clock differences in distributed processing, and characteristics of the underlying storage systems mean that documents could be processed and committed in any order.</para>
+        /// </remarks>
+        Task<WorkflowInstanceLogPage> GetLogEntriesAsync(string workflowInstanceId, long? startingTimestamp = null, int maxItems = 25, string continuationToken = null);
     }
 }
