+++ Solutions/Marain.Workflow.Storage.AzureBlob/Marain/Workflows/Storage/Internal/ContinuationToken.cs
@@ -15,7 +15,7 @@ namespace Marain.Workflows.Storage.Internal
         /// <param name="pageSize">The page size.</param>
         /// <param name="pageIndex">The next page to retrieve.</param>
         /// <param name="startingTimestamp">The starting timestamp for the log, or null to start from the first available entry.</param>
-        public ContinuationToken(int pageSize, int pageIndex, int? startingTimestamp = null)
+        public ContinuationToken(int pageSize, int pageIndex, long? startingTimestamp = null)
         {
             this.PageSize = pageSize;
             this.PageIndex = pageIndex;
@@ -35,6 +35,6 @@ namespace Marain.Workflows.Storage.Internal
         /// <summary>
         /// Gets the starting sequence number for the log, or null to start from the first available entry.
         /// </summary>
-        public int? StartingTimestamp { get; }
+        public long? StartingTimestamp { get; }
     }
 }
